+++ StabilityMatrix.Core/Models/PackageModification/InstallPackageStep.cs
@@ -8,12 +8,19 @@ public class InstallPackageStep : IPackageStep
 {
     private readonly BasePackage package;
     private readonly TorchVersion torchVersion;
+    private readonly DownloadPackageVersionOptions versionOptions;
     private readonly string installPath;
 
-    public InstallPackageStep(BasePackage package, TorchVersion torchVersion, string installPath)
+    public InstallPackageStep(
+        BasePackage package,
+        TorchVersion torchVersion,
+        DownloadPackageVersionOptions versionOptions,
+        string installPath
+    )
     {
         this.package = package;
         this.torchVersion = torchVersion;
+        this.versionOptions = versionOptions;
         this.installPath = installPath;
     }
 
@@ -25,7 +32,7 @@ public class InstallPackageStep : IPackageStep
         }
 
         await package
-            .InstallPackage(installPath, torchVersion, progress, OnConsoleOutput)
+            .InstallPackage(installPath, torchVersion, versionOptions, progress, OnConsoleOutput)
             .ConfigureAwait(false);
     }