+++ StabilityMatrix.Core/Models/Packages/A3WebUI.cs
@@ -67,9 +67,9 @@ public class A3WebUI : BaseGitPackage
             [SharedOutputType.Extras] = new[] { "outputs/extras-images" },
             [SharedOutputType.Saved] = new[] { "log/images" },
             [SharedOutputType.Img2Img] = new[] { "outputs/img2img-images" },
-            [SharedOutputType.Text2Img] = new[] { "outputs/text2img-images" },
+            [SharedOutputType.Text2Img] = new[] { "outputs/txt2img-images" },
             [SharedOutputType.Img2ImgGrids] = new[] { "outputs/img2img-grids" },
-            [SharedOutputType.Text2ImgGrids] = new[] { "outputs/text2img-grids" }
+            [SharedOutputType.Text2ImgGrids] = new[] { "outputs/txt2img-grids" }
         };
 
     [SuppressMessage("ReSharper", "ArrangeObjectCreationWhenTypeNotEvident")]
@@ -181,12 +181,11 @@ public class A3WebUI : BaseGitPackage
     public override async Task InstallPackage(
         string installLocation,
         TorchVersion torchVersion,
+        DownloadPackageVersionOptions versionOptions,
         IProgress<ProgressReport>? progress = null,
         Action<ProcessOutput>? onConsoleOutput = null
     )
     {
-        await base.InstallPackage(installLocation, torchVersion, progress).ConfigureAwait(false);
-
         progress?.Report(new ProgressReport(-1f, "Setting up venv", isIndeterminate: true));
         // Setup venv
         await using var venvRunner = new PyVenvRunner(Path.Combine(installLocation, "venv"));